pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentrx-q"
version = "0.1.0"
description = "AI-powered Reddit mod queue triage bot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "praw>=7.7.0",
    "openai>=1.0.0",
    "flask>=3.0.0",
    "sqlalchemy>=2.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "flake8>=7.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=bot --cov=dashboard --cov=database --cov-report=term-missing"

[tool.coverage.run]
omit = ["tests/*"]
